### 2021-05-06 #39 @SimonBiggs
PyMedPhys
Submitting Author: Simon Biggs (@SimonBiggs)
Package Name: PyMedPhys
One-Line Description of Package: A python library providing tools for Medical Physicists.
Repository Link (if existing): https://github.com/pymedphys/pymedphys

---

Provides a central hub for Medical Physicists to share and review each others code that improves their day-to-day job. It also provides analysis tools that supports [Medical Physics research](https://scholar.google.com/scholar?hl=en&as_sdt=0%2C5&q=pymedphys&btnG=).

See description above.

Medical Physicists.

Similar tools, `pylinac` and `pydicom`. PyMedPhys complements these packages often builds on top of them.
