### 2021-05-14 ecdc9ca
update bb diameter within config
