+++ site-specific/cancer-care-associates/config.toml
@@ -83,6 +83,9 @@ name = "rccc"
         [site.linac.directories]
         qa = '\\rccc-pdc\PExIT\Physics\RCCC Specific Files\Linac Beam Data Record\Synergy 2694'
 
+    [site.daily-wlutz]
+    bb_diameter = 12
+
 
 [[site]]
 name = "nbcc"
@@ -130,6 +133,9 @@ name = "nbcc"
 
         [site.linac.directories]
         qa = '\\rccc-pdc\PExIT\Physics\NBCCC Specific Files\Linac Beam Data Record\VersaHD 9183'
+
+    [site.daily-wlutz]
+    bb_diameter = 6.4
 
 
 [[site]]
