+++ site-specific/cancer-care-associates/config.toml
@@ -129,7 +129,7 @@ name = "nbcc"
     energies = ["6 MV", "10 MV", "6 MV FFF"]
 
         [site.linac.aliases]
-        iview = "iView"
+        iview = "9183"
 
         [site.linac.directories]
         qa = '\\rccc-pdc\PExIT\Physics\NBCCC Specific Files\Linac Beam Data Record\VersaHD 9183'
