### 2021-05-14 5f47d53
fix minimum bb diameter input
