+++ papers/gamma/paper.bib
@@ -1,0 +1,59 @@
+@article{Pearson:2017,
+  url           = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
+  archiveprefix = {arXiv},
+  author        = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
+  eprint        = {1703.04627},
+  journal       = {ArXiv e-prints},
+  keywords      = {Astrophysics - Astrophysics of Galaxies},
+  month         = mar,
+  title         = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
+  year          = 2017
+}
+
+@book{Binney:2008,
+  url       = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
+  author    = {{Binney}, J. and {Tremaine}, S.},
+  booktitle = {Galactic Dynamics: Second Edition, by James Binney and Scott Tremaine.~ISBN 978-0-691-13026-2 (HB).~Published by Princeton University Press, Princeton, NJ USA, 2008.},
+  publisher = {Princeton University Press},
+  title     = {{Galactic Dynamics: Second Edition}},
+  year      = 2008
+}
+
+@article{gaia,
+  author        = {{Gaia Collaboration}},
+  title         = {{The Gaia mission}},
+  journal       = {Astronomy and Astrophysics},
+  archiveprefix = {arXiv},
+  eprint        = {1609.04153},
+  primaryclass  = {astro-ph.IM},
+  keywords      = {space vehicles: instruments, Galaxy: structure, astrometry, parallaxes, proper motions, telescopes},
+  year          = 2016,
+  month         = nov,
+  volume        = 595,
+  doi           = {10.1051/0004-6361/201629272},
+  url           = {http://adsabs.harvard.edu/abs/2016A%26A...595A...1G}
+}
+
+@article{astropy,
+  author        = {{Astropy Collaboration}},
+  title         = {{Astropy: A community Python package for astronomy}},
+  journal       = {Astronomy and Astrophysics},
+  archiveprefix = {arXiv},
+  eprint        = {1307.6212},
+  primaryclass  = {astro-ph.IM},
+  keywords      = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
+  year          = 2013,
+  month         = oct,
+  volume        = 558,
+  doi           = {10.1051/0004-6361/201322068},
+  url           = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
+}
+
+@misc{fidgit,
+  author    = {A. Smith},
+  title     = {Fidgit: An ungodly union of GitHub and Figshare},
+  year      = {2020},
+  publisher = {GitHub},
+  journal   = {GitHub repository},
+  url       = {https://github.com/arfon/fidgit}
+}
