### 2021-05-17 60f6f25
run bipolar conversion on a 'per-beam' basis
