+++ lib/pymedphys/_monaco/delivery.py
@@ -16,7 +16,6 @@ import functools
 import re
 
 from pymedphys._imports import numpy as np
-from pymedphys._imports import streamlit as st
 
 import pymedphys._base.delivery
 import pymedphys._utilities.transforms
