### 2021-05-18 66e9dd7
update streamlit and other deps, bump version
