+++ .github/workflows/library.yml
@@ -437,6 +437,7 @@ jobs:
         if: matrix.task == 'docs' && matrix.os == 'ubuntu' && github.event_name == 'push'
         uses: JamesIves/github-pages-deploy-action@4.1.3
         with:
+          GITHUB_TOKEN: ${{ secrets.PYMEDPHYS_BOT_TOKEN }}
           BRANCH: docs # The branch the action should deploy to.
           FOLDER: lib/pymedphys/docs/_build/html # The folder the action should deploy.
           CLEAN: true # Automatically remove deleted files from the deploy branch
