### 2021-05-18 efae518
change docs deploy to use ssh instead
