### 2021-05-21 #39 @lwasser
hi there @SimonBiggs ! Welcome to pyopensci!
Can you  tell us a bit more about this submission? It doesn't look to me like this is a python package on it's own, rather a community / subset of packages similar maybe to astropy? please let me know what you are looking for here. And perhaps you can provide some more information about the peer review part of things as that is what we do here at pyopensci!
