### 2021-05-22 #39 @SimonBiggs
Hi @lwasser,

We're in an interesting scenario where people are using the code within publications:

* https://docs.pymedphys.com/lib/ref/gamma.html
* https://scholar.google.com/scholar?hl=en&as_sdt=0%2C5&q=pymedphys+gamma&btnG=

But we don't actually have a paper out there. I began to attempt to write a JOSS paper (https://joss.theoj.org/) for the gamma implementation, but it appears that one repo can't have multiple papers, it'd be handy to be able to have just a "pymedphys.gamma" paper, and then have ways where other contributors can spin off papers/get peer review, for sub-modules.

I saw that for a JOSS submission it can be easier if it has gone through the `pyOpenSci` process first. Might it be possible to attempt to have the above `pymedphys.gamma` module to go through review? And then see how easy it would be to submit a JOSS paper just for that subsection of PyMedPhys? This would then pave the way for others in the community to follow a similar process...

Cheers,
Simon
