+++ lib/pymedphys/tests/experimental/pseudonymisation/test_pseudonymisation.py
@@ -18,11 +18,12 @@ from pymedphys._dicom.anonymise import (
 )
 from pymedphys._dicom.constants.core import DICOM_SOP_CLASS_NAMES_MODE_PREFIXES
 from pymedphys._dicom.utilities import remove_file
-from pymedphys.experimental import pseudonymisation as pseudonymisation_api
 from pymedphys.tests.dicom.test_anonymise import (
     dicom_dataset_from_dict,
     get_test_filepaths,
 )
+
+from pymedphys._experimental.pseudonymisation import api as pseudonymisation_api
 
 
 @pytest.mark.pydicom
