### 2021-05-31 1d39f84
replace previous API with a stub and error message that directs users to the issue. Remove from docs.
