+++ lib/pymedphys/docs/lib/ref/experimental/index.rst
@@ -15,4 +15,3 @@ Experimental Modules
     :maxdepth: 1
 
     pinnacle
-    pseudonymisation
