### 2021-05-31 daac38f
unexpose the CLI
