### 2021-06-02 #39 @lwasser
hi again @SimonBiggs help me understand this a bit better. we currently review python packages which would follow the JOSS model that there is one package within a single repo that is curated, citable etc. It sounds as if this may be a set of submodules which similar to JOSS, we don't have the infrastructure right now to support review for as the citation would be on the repo rather than sub sections of the repo. Some issues would come up because:

*  any archive of this would be for the entire repo yet it sounds like the review would only be for a subset of the repo
* how do you track citation for a submodule within a larger repo - tools like zenodo are repo focused not submodule focused
* tracking maintenance over time of a sub module would get challenging.

Do i understand this correctly? We may have to put this on hold if that is the case for the time being until we figure out a way that this type of review could be implemented in a sustainable way. But please do let me know if i am mis understanding the goal here.
