### 2021-06-02 #39 @lwasser
i will say that we are planning to work with other communities to implement uniform standards and such - pangeo being one of them - so perhaps there is a way we can work together in the future even if we can't perform a review right now.
