### 2021-06-02 #39 @SimonBiggs
> Do i understand this correctly? We may have to put this on hold if that is the case for the time being until we figure out a way that this type of review could be implemented in a sustainable way.

Yup, I think you've got it to a tee. I'm all for making sure which ever way it occurs it is sustainable. Happy to just stay on the radar for now until the approach is more supported.

Cheers :slightly_smiling_face:,
Simon
