+++ requirements.txt
@@ -115,7 +115,7 @@ tqdm==4.60.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (py
 traitlets==5.0.5; python_full_version >= "3.6.1" and python_version >= "3.7" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7")
 typing-extensions==3.10.0.0
 tzlocal==2.1; python_version >= "3.6"
-urllib3==1.26.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
+urllib3==1.26.5; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
 validators==0.18.2; python_version >= "3.6"
 watchdog==2.1.1; python_version >= "3.6"
 wcwidth==0.2.5; python_full_version >= "3.6.1" and python_version >= "3.7"
